Added test_main.py
