+++ src/network/client.py
@@ -10,18 +10,21 @@ class Client(object):
     def __init__(self):
         pass
 
+    @staticmethod
     def generate_keys():
         modulus_lenght = 256 * 4
         private_key = RSA.generate(modulus_lenght, Random.new().read)
         public_key = private_key.publickey()
         return private_key, public_key
 
+    @staticmethod
     def encrypt_private_key(a_message, private_key):
         encryptor = Salsa20.new(private_key)
         encrypted_msg = encryptor.encrypt(a_message)
         encoded_encrypted_msg = base64.b64encode(encrypted_msg)
         return encoded_encrypted_msg
 
+    @staticmethod
     def decrypt_public_key(encoded_encrypted_msg, public_key):
         encryptor = Salsa20.new(public_key)
         decoded_encrypted_msg = base64.b64decode(encoded_encrypted_msg)
